// src/shared/enums/permissions.enum.ts
export enum Permissions {
  MANAGE_USERS = 'manage_users',
  MANAGE_CATALOGS = 'manage_catalogs',
  UPDATE_DESTINATIONS = 'update_destinations',
  CONFIGURE_ITINERARY = 'configure_itinerary',
  PUBLISH_TOURS = 'publish_tours',
  VIEW_RESERVATIONS = 'view_reservations',
  VIEW_REPORTS = 'view_reports',
  ACCESS_LOGS = 'access_logs',
  VIEW_INFORMATION = 'view_information',
}